+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/graph/IrBinding.kt
@@ -7,11 +7,11 @@ import dev.zacsweers.metro.compiler.MetroAnnotations
 import dev.zacsweers.metro.compiler.appendLineWithUnderlinedContent
 import dev.zacsweers.metro.compiler.capitalizeUS
 import dev.zacsweers.metro.compiler.expectAs
-import dev.zacsweers.metro.compiler.expectAsOrNull
 import dev.zacsweers.metro.compiler.graph.BaseBinding
 import dev.zacsweers.metro.compiler.graph.LocationDiagnostic
 import dev.zacsweers.metro.compiler.ir.BindsCallable
 import dev.zacsweers.metro.compiler.ir.ClassFactory
+import dev.zacsweers.metro.compiler.ir.Format
 import dev.zacsweers.metro.compiler.ir.IrAnnotation
 import dev.zacsweers.metro.compiler.ir.IrContextualTypeKey
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
@@ -30,9 +30,9 @@ import dev.zacsweers.metro.compiler.ir.parameters.Parameters
 import dev.zacsweers.metro.compiler.ir.rawType
 import dev.zacsweers.metro.compiler.ir.regularParameters
 import dev.zacsweers.metro.compiler.ir.render
+import dev.zacsweers.metro.compiler.ir.renderForDiagnostic
 import dev.zacsweers.metro.compiler.ir.reportableDeclaration
 import dev.zacsweers.metro.compiler.ir.requireSimpleType
-import dev.zacsweers.metro.compiler.ir.sourceGraphIfMetroGraph
 import dev.zacsweers.metro.compiler.isWordPrefixRegex
 import dev.zacsweers.metro.compiler.memoize
 import dev.zacsweers.metro.compiler.reportCompilerBug
@@ -42,18 +42,16 @@ import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationContainer
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationParent
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationWithName
-import org.jetbrains.kotlin.ir.declarations.IrField
 import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.declarations.IrPackageFragment
 import org.jetbrains.kotlin.ir.declarations.IrProperty
 import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
 import org.jetbrains.kotlin.ir.declarations.IrValueParameter
+import org.jetbrains.kotlin.ir.types.IrErrorType
 import org.jetbrains.kotlin.ir.types.IrType
-import org.jetbrains.kotlin.ir.types.isUnit
 import org.jetbrains.kotlin.ir.types.typeOrFail
 import org.jetbrains.kotlin.ir.util.callableId
 import org.jetbrains.kotlin.ir.util.classId
-import org.jetbrains.kotlin.ir.util.dumpKotlinLike
 import org.jetbrains.kotlin.ir.util.hasAnnotation
 import org.jetbrains.kotlin.ir.util.isPropertyAccessor
 import org.jetbrains.kotlin.ir.util.kotlinFqName
@@ -63,7 +61,6 @@ import org.jetbrains.kotlin.ir.util.parentDeclarationsWithSelf
 import org.jetbrains.kotlin.ir.util.propertyIfAccessor
 import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.ClassId
-import org.jetbrains.kotlin.name.Name
 
 internal sealed interface IrBinding : BaseBinding<IrType, IrTypeKey, IrContextualTypeKey> {
   override val typeKey: IrTypeKey
@@ -215,6 +212,13 @@ internal sealed interface IrBinding : BaseBinding<IrType, IrTypeKey, IrContextua
     override val contextualTypeKey: IrContextualTypeKey,
     override val parameters: Parameters,
   ) : StaticBinding {
+
+    init {
+      if (contextualTypeKey.typeKey.type is IrErrorType) {
+        error("wtf")
+      }
+    }
+
     override val dependencies: List<IrContextualTypeKey> by memoize {
       parameters.allParameters.map { it.contextualTypeKey }
     }
@@ -855,199 +859,6 @@ internal sealed interface IrBinding : BaseBinding<IrType, IrTypeKey, IrContextua
       }
 
     override fun toString() = renderDescriptionDiagnostic(short = true, underlineTypeKey = false)
-  }
-}
-
-context(builder: StringBuilder)
-private fun IrClass.renderForDiagnostic(
-  short: Boolean,
-  annotations: MetroAnnotations<IrAnnotation>,
-  underlineTypeKey: Boolean,
-) {
-  with(builder) {
-    renderAnnotations(annotations, short, isClass = false)
-    append(kind.codeRepresentation)
-    append(' ')
-    if (underlineTypeKey) {
-      appendLineWithUnderlinedContent(name.asString())
-    } else {
-      append(name.asString())
-    }
-  }
-}
-
-private enum class Format {
-  DECLARATION,
-  CALL;
-
-  val isDeclaration: Boolean
-    get() = this == DECLARATION
-
-  val isCall: Boolean
-    get() = this == CALL
-}
-
-private fun StringBuilder.renderForDiagnostic(
-  declaration: IrDeclarationParent,
-  short: Boolean,
-  typeKey: IrTypeKey,
-  annotations: MetroAnnotations<IrAnnotation>?,
-  parameters: Parameters,
-  isProperty: Boolean?,
-  underlineTypeKey: Boolean,
-  format: Format = Format.DECLARATION,
-) {
-  val property: IrProperty?
-  val name: Name
-  val type: IrType
-  when (declaration) {
-    is IrField -> {
-      property = null
-      name = declaration.name
-      type = declaration.type
-    }
-    is IrFunction -> {
-      property = declaration.propertyIfAccessor.expectAsOrNull<IrProperty>()
-      name = (property ?: declaration).name
-      type = declaration.returnType
-    }
-    is IrProperty -> {
-      property = declaration
-      name = declaration.name
-      type =
-        declaration.getter?.returnType
-          ?: declaration.backingField?.type
-          ?: reportCompilerBug("No getter or backing field")
-    }
-    else -> {
-      reportCompilerBug("Unsupported declaration type: ${declaration.dumpKotlinLike()}")
-    }
-  }
-
-  val isProperty = isProperty == true || property != null
-
-  if (format.isDeclaration) {
-    annotations?.let { renderAnnotations(it, short, isClass = false) }
-    if (isProperty) {
-      if (property != null) {
-        if (property.isVar) {
-          if (property.isLateinit) {
-            append("lateinit ")
-          }
-          append("var ")
-        } else {
-          append("val ")
-        }
-      } else {
-        append("val ")
-      }
-    } else {
-      append("fun ")
-    }
-
-    if (parameters.contextParameters.isNotEmpty()) {
-      parameters.contextParameters.joinTo(this, ", ", prefix = "context(", postfix = ")\n") {
-        it.name.asString() + ": " + it.typeKey.render(short = short)
-      }
-    }
-  }
-
-  val dispatchReceiverName =
-    declaration.parentClassOrNull?.sourceGraphIfMetroGraph?.name?.asString()
-  var hasReceiver = false
-
-  if (format.isCall) {
-    dispatchReceiverName?.let {
-      append(it)
-      hasReceiver = true
-    }
-  }
-
-  parameters.extensionReceiverParameter?.let {
-    if (format.isCall) {
-      // Put the receiver in parens for context
-      append('(')
-    }
-    it.typeKey.qualifier?.let { qualifier ->
-      append(qualifier.render(short = short, "receiver"))
-      append(' ')
-    }
-    append(it.typeKey.render(short = short))
-    if (format.isCall) {
-      // Put the receiver in parens for context
-      append(')')
-    }
-    hasReceiver = true
-  }
-
-  if (hasReceiver) {
-    append('.')
-  }
-
-  append(name.asString())
-
-  val paramsToDisplay =
-    if (format.isCall) {
-      // Likely member inject() call
-      parameters.regularParameters.filterNot { it.isAssisted }
-    } else {
-      parameters.regularParameters
-    }
-  if (paramsToDisplay.isNotEmpty()) {
-    paramsToDisplay.joinTo(this, ", ", prefix = "(", postfix = ")\n") {
-      it.name.asString() + ": " + it.typeKey.render(short = short, includeQualifier = true)
-    }
-  } else if (!isProperty) {
-    append("()")
-  }
-
-  if (!(declaration is IrFunction && type.isUnit())) {
-    append(": ")
-    val returnTypeString = typeKey.render(short = short, includeQualifier = false)
-    if (underlineTypeKey) {
-      appendLineWithUnderlinedContent(returnTypeString)
-    } else {
-      append(returnTypeString)
-    }
-  }
-}
-
-private fun StringBuilder.renderAnnotations(
-  annotations: MetroAnnotations<IrAnnotation>,
-  short: Boolean,
-  isClass: Boolean,
-) {
-  val annotationStrings =
-    with(annotations) {
-      buildList {
-        qualifier?.let { add(it.render(short = short)) }
-        if (isBinds) add("@Binds")
-        if (isProvides) add("@Provides")
-        if (isIntoSet) add("@IntoSet")
-        if (isElementsIntoSet) add("@ElementsIntoSet")
-        if (isMultibinds) add("@Multibinds")
-        if (isBindsOptionalOf) add("@BindsOptionalOf")
-        scope?.let { add(it.render(short = short)) }
-        if (isIntoMap) add("@IntoMap")
-        mapKeys.forEach { add(it.render(short = short)) }
-        if (isClass) {
-          if (isInject) add("@Inject")
-        }
-      }
-    }
-  when (annotationStrings.size) {
-    0 -> {
-      // do nothing
-    }
-    1,
-    2 -> {
-      annotationStrings.joinTo(this, " ")
-      append(' ')
-    }
-    else -> {
-      annotationStrings.joinTo(this, "\n")
-      appendLine()
-    }
   }
 }